+++ DirectQ/net_main.cpp
@@ -37,7 +37,7 @@ static bool	listening = false;
 bool	slistInProgress = false;
 bool	slistSilent = false;
 bool	slistLocal = true;
-static double	slistStartTime;
+static float	slistStartTime;
 static int		slistLastShown;
 
 static void Slist_Send(void *soak);
@@ -66,9 +66,9 @@ bool	configRestored = false;
 int	net_driverlevel;
 
 
-double			net_time;
+float			net_time;
 
-double SetNetTime(void)
+float SetNetTime(void)
 {
 	net_time = Sys_FloatTime();
 	return net_time;
@@ -425,7 +425,7 @@ NET_CheckNewConnections
 
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 } vcrConnect;
@@ -485,7 +485,7 @@ returns -1 if connection is invalid
 
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 	int		ret;
@@ -551,7 +551,7 @@ returns -1 if the connection died
 */
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 	int		r;
@@ -629,7 +629,7 @@ bool NET_CanSendMessage (qsocket_t *sock)
 
 int NET_SendToAll(sizebuf_t *data, int blocktime)
 {
-	double		start;
+	float		start;
 	int			i;
 	int			count = 0;
 	bool	state1 [MAX_SCOREBOARD];
@@ -822,7 +822,7 @@ void NET_Poll(void)
 }
 
 
-void SchedulePollProcedure(PollProcedure *proc, double timeOffset)
+void SchedulePollProcedure(PollProcedure *proc, float timeOffset)
 {
 	PollProcedure *pp, *prev;
