+++ DirectQ/server.h
@@ -39,10 +39,10 @@ typedef struct
 	bool	paused;
 	bool	loadgame;			// handle connections specially
 
-	double		time;
+	float		time;
 	
 	int			lastcheck;			// used by PF_checkclient
-	double		lastchecktime;
+	float		lastchecktime;
 	
 	char		name[64];			// map name
 
@@ -86,7 +86,7 @@ typedef struct client_s
 	bool		privileged;			// can execute any host command
 	bool		sendsignon;			// only valid before spawned
 
-	double			last_message;		// reliable messages must be sent
+	float			last_message;		// reliable messages must be sent
 										// periodically
 
 	struct qsocket_s *netconnection;	// communications handle
@@ -212,7 +212,7 @@ extern	client_t	*host_client;
 
 extern	jmp_buf 	host_abortserver;
 
-extern	double		host_time;
+extern	float		host_time;
 
 extern	edict_t		*sv_player;
